% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_tstim.R
\name{generate_tstim_vect}
\alias{generate_tstim_vect}
\title{Generate timing information within each trial for SMI eye tracking data}
\usage{
generate_tstim_vect(trial_df)
}
\arguments{
\item{trial_df}{A data frame for each trial in the SMI eye tracking data}
}
\value{
A data frame that has a t.stim timing column
}
\description{
Generate timing information within each trial for SMI eye tracking data
}
