% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_smi.R
\name{preprocess_smi}
\alias{preprocess_smi}
\title{Preprocess SMI eye tracking data file}
\usage{
preprocess_smi(d, x_max = 1680, y_max = 1050, samp_rate = 120,
  avg_eyes = TRUE)
}
\arguments{
\item{d}{A data frame with eye tracking data.}

\item{x_max}{A number indicating the max pixel dimension in the horizontal plane.}

\item{y_max}{A number indicating the max pixel dimension in the vertical plane.}

\item{samp_rate}{A number indicating the sampling rate of the eye tracker that generated the data.}

\item{avg_eyes}{Logical indicating whether data from each eye should be averaged together.}
}
\value{
A processed data frame that is ready for analysis.
}
\description{
Preprocess SMI eye tracking data file
}
\examples{
preprocess_smi(d = df, x_max = 1680, y_max = 1050, samp_rate = 120, avg_eyes = TRUE)
}
