% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_smi.R
\name{read_smi}
\alias{read_smi}
\title{Read in data file created by the SMI tracker's IDF converter utility}
\usage{
read_smi(file_name, file_path, length_header = 40)
}
\arguments{
\item{file_name}{A string with the name of the raw data file.}

\item{file_path}{A string with the path to the raw data file.}

\item{length_header}{A number indicating the number of lines in the header of the raw data file.}
}
\value{
A data frame ready for preprocessing
}
\description{
Read in data file created by the SMI tracker's IDF converter utility
}
\examples{
read_smi(file_name = "eye_data_idf.txt", file_path = "../data/", length_header = 40)
}
