% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_smi.R
\name{score_smi}
\alias{score_smi}
\title{Score trial for SMI eye tracking data}
\usage{
score_smi(d, trial_column = trial_num, response_column)
}
\arguments{
\item{d}{A data frame with eye tracking data}

\item{trial_column}{A string indicating the column name that tracks trial number.}

\item{response_column}{A string indicating the column name that tracks where participants were looking.}
}
\value{
A data frame that has an RT for the first gaze shift in each trial.
}
\description{
Score trial for SMI eye tracking data
}
\examples{
score_smi(d = df, trial_column = trial_num, response_column = ss_looking_char)
}
