% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_n.R
\name{to.n}
\alias{to.n}
\title{Convert variable to numeric}
\usage{
to.n(x)
}
\arguments{
\item{x}{A string to be converted}
}
\value{
A numeric
}
\description{
Convert variable to numeric
}
\examples{
to_n("10")
}
